+++ app/(main)/dashboard/page.tsx
@@ -9,6 +9,7 @@ import Calendar from '@/app/(main)/dashboard/components/Calendar'
 // import SearchInput from '@/components/common/Input/SearchInput'
 // import Modal from '@/components/common/Modal'
 // import { ModalMessageType } from '@/components/common/Modal/types'
+import SignOutButton from '@/components/common/SignOutButton'
 import HeaderSkeleton from '@/components/common/Skeleton/HeaderSkeleton'
 import Map from '@/components/domain/map/Map'
 // import { useSearchSingleVehicle } from '@/hooks/useVehicleLocationSearch'
@@ -216,8 +217,11 @@ const DashboardPage = () => {
                 </div>
 
                 <NoticeListBoard />
-            </section>
 
+                <div className={styles.signOutButton}>
+                    <SignOutButton />
+                </div>
+            </section>
             {/* <Modal
                 isOpen={isOpen}
                 message={modalMessage as ModalMessageType}
