+++ components/common/Button/BackButton/index.tsx
@@ -2,6 +2,8 @@
 
 import { useRouter } from 'next/navigation'
 
+import { BackIcon } from '@/public/icons'
+
 import * as styles from './styles.css'
 
 interface BackButtonProps {
@@ -15,7 +17,7 @@ const BackButton = ({ onBackButtonClick, className }: BackButtonProps) => {
 
     return (
         <button onClick={navigateBeforePage} className={`${styles.button} ${className ?? ''}`}>
-            -
+            <BackIcon style={{ width: '24px', height: '24px' }} />
         </button>
     )
 }
